+++ R/sig_fit.R
@@ -51,6 +51,10 @@
 #'   H_infer
 #'   H
 #'
+#'   ## Get clusters/groups
+#'   z <- get_groups(H_infer, method = "k-means")
+#'   show_groups(z)
+#'
 #'   H_dt <- sig_fit(V, W, return_class = "data.table")
 #'   H_dt
 #'