+++ NEWS.md
@@ -1,5 +1,8 @@
 # sigminer 1.0.6
 
+- Added `show_groups()` to show the signature contribution in each group from `get_groups()`.
+- Expanded clustering in `get_groups()` to result of `sig_fit()`.
+- Properly handled null-count samples in `sig_fit_bootstrap_batch()`.
 - Added strand bias labeling for INDEL.
 - Added COSMIC TSB signatures.